moving model_helper functions to text_utils and test_utils
